% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_se.R
\name{load_se}
\alias{load_se}
\title{Load summarizedExperiment}
\usage{
load_se()
}
\value{
A summarizedExperiment object containing the merged data from multiple summarizedExperiment objects. It contains cosmic and depmap data on cell lines.
}
\description{
This function combines multiple summarizedExperiment objects into a single object.
}
\details{
This function is used to merge multiple summarizedExperiment objects into a single object by row-binding them together.
}
\examples{
# Example usage
se <- load_se()

}
