% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap}
\usage{
plotHeatmap(
  se,
  mutations,
  genes = NULL,
  group_var = NULL,
  group_lvl = NULL,
  cell_lines = NULL,
  remove_na = TRUE
)
}
\arguments{
\item{se}{A summarizedExperiment object containing mutation data.}

\item{mutations}{A character vector specifying the mutations of interest.}

\item{genes}{A character vector specifying the genes of interest. Default is NULL.}

\item{group_var}{A variable (column) name included in colData(se). Can be used to subset the dataset.}

\item{group_lvl}{A level present in group_var. Only cell lines that are included in this subgroup are taken into account. Required if group_var is provided.}

\item{cell_lines}{A character vector specifying the cell lines to include in the plot. Default is NULL, which means that all cell lines will be included (except if group_Var/group_lvl is specified)}

\item{remove_na}{A logical value indicating whether to remove cell lines that have no data on mutations. Default is set to TRUE}
}
\value{
Plots oncoplots showing which mutations are present in cell lines and optionally also gene expression data. In addition,  list containing the mutations and expression data if genes is provided. Otherwise, returns the mutations.
}
\description{
This function generates a heatmap plot based on mutation data and gene expression data from cell lines.
}
\examples{
plotHeatmap(se, mutations = c("APC", "KRAS"), genes = c("TP53", "CDK9"),
group_var = "oncoTreeCode", group_lvl = "COAD")
}
