% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCellLine.R
\name{returnCellLines}
\alias{returnCellLines}
\title{Return Cell Lines}
\usage{
returnCellLines(
  se,
  mutations,
  genes = NULL,
  group_var = NULL,
  group_lvl = NULL
)
}
\arguments{
\item{se}{The summarizedExperiment that comes with the package.}

\item{mutations}{A character vector specifying the mutations of interest.}

\item{genes}{A character vector specifying the genes of interest. In case of NULL, no gene expression data is shown.}

\item{group_var}{A character string specifying the grouping variable in the colData of \code{se}. Default is NULL.}

\item{group_lvl}{A character string specifying the level of the grouping variable. Required if \code{group_var} is provided.}
}
\value{
A data frame with cell line names and gene expression values.
}
\description{
This function returns a data frame containing cell lines and gene expression values for the specified mutations and genes.
}
\examples{
\dontrun{
returnCellLines(se, mutations = C('SMAD4', 'APC'), genes = c('CDK2', 'PTEN'), group_var = "oncoTreeCode", group_lvl = "COAD")
}



}
