% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se2}
\alias{se2}
\title{SummarizedExperiment object with depmap and cosmic data.}
\format{
\subsection{\code{se2}}{

A summarizedExperiment object
}
}
\source{
www.depmap.org, https://cancer.sanger.ac.uk/cell_lines, https://www.cancerrxgene.org/celllines
}
\usage{
se2
}
\description{
The SE object stores data about mutations and gene expression data collected from cosmic and
depmap. Part 2/2
}
\keyword{datasets}
